% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square.R
\name{square}
\alias{square}
\title{Square a number}
\usage{
square(x)
}
\arguments{
\item{x}{number to be squared}
}
\value{
squared number
}
\description{
Square Function
}
\details{
This function is designed to square a function
}
\examples{
square(2)
}
\author{
Caleb
}

