% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squareC.R
\name{squareC}
\alias{squareC}
\title{Square a number}
\usage{
squareC(x)
}
\arguments{
\item{x}{number to be squared}
}
\value{
squared number
}
\description{
Square Function
}
\details{
This function is designed to square a function, but this function will Call C
}
\examples{
squareC(2)
}
\author{
Caleb
}

